Move Button, Knob & Popup into Controls.cpp,h
